+++ src/Phan/Analysis/ConditionVisitor/BinaryCondition.php
@@ -28,10 +28,18 @@ interface BinaryCondition
     public function analyzeClassCheck(ConditionVisitorInterface $visitor, $object, $expr) : Context;
 
     /**
-     * Assert that this condition applies to the function call $call_node (i.e. is_string($object) OPERATION $expr)
+     * Assert that this condition applies to the function call $call_node (e.g. is_string($object) OPERATION $expr)
      *
      * @param Node $call_node a node of kind AST_CALL
      * @param Node|string|int|float $expr
      */
     public function analyzeCall(ConditionVisitorInterface $visitor, Node $call_node, $expr) : ?Context;
+
+    /**
+     * Assert that this condition applies to the function call $complex_node (e.g. ($var instanceof Xyz) OPERATION $expr)
+     *
+     * @param Node $complex_node a node of various kinds
+     * @param Node|string|int|float $expr a node with a constant value
+     */
+    public function analyzeComplexCondition(ConditionVisitorInterface $visitor, Node $complex_node, $expr) : ?Context;
 }