+++ src/Phan/Analysis/ConditionVisitor/IdenticalCondition.php
@@ -56,4 +56,24 @@ class IdenticalCondition implements BinaryCondition
             return (new NegatedConditionVisitor($code_base, $context))->visitCall($call_node);
         }
     }
+
+    public function analyzeComplexCondition(ConditionVisitorInterface $visitor, Node $complex_node, $expr) : ?Context
+    {
+        if (!$expr instanceof Node) {
+            return null;
+        }
+        $code_base = $visitor->getCodeBase();
+        $context = $visitor->getContext();
+        $value = UnionTypeVisitor::unionTypeFromNode($code_base, $context, $expr)->asSingleScalarValueOrNullOrSelf();
+        if (!\is_bool($value)) {
+            return null;
+        }
+        if ($value) {
+            // e.g. `if (is_string($x) === true)`
+            return (new ConditionVisitor($code_base, $context))->__invoke($complex_node);
+        } else {
+            // e.g. `if (is_string($x) === false)`
+            return (new NegatedConditionVisitor($code_base, $context))->__invoke($complex_node);
+        }
+    }
 }